created folder move-generation and moved move generation handler there
