+++ angular-chess/src/app/utils/move-generation/move-generation.queen.handler.ts
@@ -1,22 +1,30 @@
-import { Board, Position } from "../types/board.t";
-import { Move, Piece, PieceType } from "../types/pieces.t";
-import BoardUtils from "../utils/board.utils";
-import PositionUtils from "../utils/position.utils";
+import { Board, Position } from "src/app/types/board.t";
+import { Move, Piece, PieceType } from "src/app/types/pieces.t";
+import BoardUtils from "../board.utils";
+import PositionUtils from "../position.utils";
 import { MoveGenerationHandler } from "./move-generation.handler";
 
-export class MoveGenerationBishopHandler implements MoveGenerationHandler {
+export class MoveGenerationQueenHandler implements MoveGenerationHandler {
 
   public canHandle(piece: Piece): boolean {
-    return piece.type === PieceType.BISHOP;
+    return piece.type === PieceType.QUEEN;
   }
 
   public getMoves(piece: Piece, board: Board): Move[] {
+    const frontSquares: Position[] = BoardUtils.getFreeFrontSquares(board, piece, 8 - piece.position.row);
+    const backSquares: Position[] = BoardUtils.getFreeBackSquares(board, piece, piece.position.row - 1);
+    const leftSquares: Position[] = BoardUtils.getFreeLeftSquares(board, piece, piece.position.column - 1);
+    const rightSquares: Position[] = BoardUtils.getFreeRightSquares(board, piece, 8 - piece.position.column);
     const frontLeftSquares: Position[] = BoardUtils.getFreeFrontLeftSquares(board, piece, Math.min(8 - piece.position.row, piece.position.column - 1));
     const frontRightSquares: Position[] = BoardUtils.getFreeFrontRightSquares(board, piece, Math.min(8 - piece.position.row, 8 - piece.position.column));
     const backLeftSquares: Position[] = BoardUtils.getFreeBackLeftSquares(board, piece, Math.min(piece.position.row - 1, piece.position.column - 1));
     const backRightSquares: Position[] = BoardUtils.getFreeBackRightSquares(board, piece, Math.min(piece.position.row - 1, 8 - piece.position.column));
 
     const fieldsToMove = [
+      ...frontSquares,
+      ...backSquares,
+      ...leftSquares,
+      ...rightSquares,
       ...frontLeftSquares,
       ...frontRightSquares,
       ...backLeftSquares,
@@ -32,12 +40,20 @@ export class MoveGenerationBishopHandler implements MoveGenerationHandler {
   }
 
   public getCaptures(piece: Piece, board: Board): Move[] {
+    const frontSquares: Position[] = BoardUtils.getOccupiedFrontSquare(board, piece, 8 - piece.position.row);
+    const backSquares: Position[] = BoardUtils.getOccupiedBackSquare(board, piece, piece.position.row - 1);
+    const leftSquares: Position[] = BoardUtils.getOccupiedLeftSquare(board, piece, piece.position.column - 1);
+    const rightSquares: Position[] = BoardUtils.getOccupiedRightSquare(board, piece, 8 - piece.position.column);
     const frontLeftSquares: Position[] = BoardUtils.getOccupiedFrontLeftSquare(board, piece, Math.min(8 - piece.position.row, piece.position.column - 1));
     const frontRightSquares: Position[] = BoardUtils.getOccupiedFrontRightSquare(board, piece, Math.min(8 - piece.position.row, 8 - piece.position.column));
     const backLeftSquares: Position[] = BoardUtils.getOccupiedBackLeftSquare(board, piece, Math.min(piece.position.row - 1, piece.position.column - 1));
     const backRightSquares: Position[] = BoardUtils.getOccupiedBackRightSquare(board, piece, Math.min(piece.position.row - 1, 8 - piece.position.column));
 
     const fieldsToMove = [
+      ...frontSquares,
+      ...backSquares,
+      ...leftSquares,
+      ...rightSquares,
       ...frontLeftSquares,
       ...frontRightSquares,
       ...backLeftSquares,
